+++ .gitignore
@@ -6,6 +6,7 @@
 # Tool Runtime Dir
 /[Tt]ools/
 /.dotnet/
+/.packages/
 
 # User-specific files
 *.suo
