+++ .vsts-dotnet-ci.yml
@@ -11,42 +11,42 @@ resources:
     image: mcr.microsoft.com/dotnet-buildtools/prereqs:ubuntu-16.04-mlnet-20200515184230-2c829e8
 
 jobs:
-# - template: /build/ci/job-template.yml
-#   parameters:
-#     name: Centos_x64_NetCoreApp31
-#     buildScript: ./build.sh
-#     container: CentosContainer
-#     customMatrixes:
-#       Debug_Build:
-#         _configuration: Debug-netcoreapp3_1
-#         _config_short: DI
-#         _includeBenchmarkData: false
-#         _targetFramework: netcoreapp3.1
-#       Release_Build:
-#         _configuration: Release-netcoreapp3_1
-#         _config_short: RI
-#         _includeBenchmarkData: true
-#         _targetFramework: netcoreapp3.1
-#     innerLoop: true
-#     pool:
-#       name:  Hosted Ubuntu 1604
+- template: /build/ci/job-template.yml
+  parameters:
+    name: Centos_x64_NetCoreApp31
+    buildScript: ./build.sh
+    container: CentosContainer
+    customMatrixes:
+      Debug_Build:
+        _configuration: Debug-netcoreapp3_1
+        _config_short: DI
+        _includeBenchmarkData: false
+        _targetFramework: netcoreapp3.1
+      Release_Build:
+        _configuration: Release-netcoreapp3_1
+        _config_short: RI
+        _includeBenchmarkData: true
+        _targetFramework: netcoreapp3.1
+    innerLoop: true
+    pool:
+      name: Hosted Ubuntu 1604
 
-# - template: /build/ci/job-template.yml
-#   parameters:
-#     name: Ubuntu_x64_NetCoreApp21
-#     buildScript: ./build.sh
-#     container: UbuntuContainer
-#     innerLoop: true
-#     pool:
-#       name:  Hosted Ubuntu 1604
+- template: /build/ci/job-template.yml
+  parameters:
+    name: Ubuntu_x64_NetCoreApp21
+    buildScript: ./build.sh
+    container: UbuntuContainer
+    innerLoop: true
+    pool:
+      name: Hosted Ubuntu 1604
 
-# - template: /build/ci/job-template.yml
-#   parameters:
-#     name: MacOS_x64_NetCoreApp21
-#     buildScript: ./build.sh
-#     innerLoop: true
-#     pool:
-#       name: Hosted macOS
+- template: /build/ci/job-template.yml
+  parameters:
+    name: MacOS_x64_NetCoreApp21
+    buildScript: ./build.sh
+    innerLoop: true
+    pool:
+      name: Hosted macOS
 
 - template: /build/ci/job-template.yml
   parameters:
