+++ src/Redist/Microsoft.ML.DnnImageFeaturizer.ModelRedist/Microsoft.ML.DnnImageFeaturizer.ModelRedist.proj
@@ -123,4 +123,6 @@
 
   <Target Name="Test" />
 
+  <Target Name="PerformanceTest"/>
+
 </Project>
