+++ test/Microsoft.ML.PerformanceTests/Harness/Metrics.cs
@@ -11,7 +11,7 @@ using BenchmarkDotNet.Columns;
 using BenchmarkDotNet.Reports;
 using BenchmarkDotNet.Running;
 
-namespace Microsoft.ML.Benchmarks
+namespace Microsoft.ML.PerformanceTests
 {
     public abstract class WithExtraMetrics : BenchmarkBase
     {
