+++ test/Microsoft.ML.PerformanceTests/Helpers/EnvironmentFactory.cs
@@ -7,7 +7,7 @@ using Microsoft.ML.Runtime;
 using Microsoft.ML.Trainers;
 using Microsoft.ML.Transforms;
 
-namespace Microsoft.ML.Benchmarks
+namespace Microsoft.ML.PerformanceTests
 {
     internal static class EnvironmentFactory
     {
