+++ test/Microsoft.ML.PerformanceTests/Microsoft.ML.PerformanceTests.csproj
@@ -2,6 +2,8 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <GenerateProgramFile>false</GenerateProgramFile>
+    <IsUnitTestProject>false</IsUnitTestProject>
+    <IsPerformanceTestProject>true</IsPerformanceTestProject>
   </PropertyGroup>
   <ItemGroup>
     <Compile Remove="BenchmarkDotNet.Artifacts\**" />
@@ -30,4 +32,20 @@
     <NativeAssemblyReference Include="MklImports" />
     <NativeAssemblyReference Condition="'$(OS)' == 'Windows_NT'" Include="libiomp5md" />
   </ItemGroup>
+  <ItemGroup>
+    <MetaFile Update="@(MetaFile)">
+      <Url>https://aka.ms/mlnet-resources/meta/%(Identity)</Url>
+      <DestinationFile>$([System.IO.Path]::GetTempPath())/MLNET/</DestinationFile>
+    </MetaFile>
+    <TensorflowMetaFile Include="@(MetaFile->'$([System.IO.Path]::GetTempPath())/MLNET/%(Identity)')" />
+  </ItemGroup>
+  <Target Name="DownloadTensorflowMetaFiles" Inputs="@(TensorflowMetaFile)" Outputs="%(TensorflowMetaFile.DestinationFile)">
+    <Message Importance="High" Text="Downloading tensorflow meta files... %(TensorflowMetaFile.DestinationFile)" />
+    <DownloadFile
+            SourceUrl="%(MetaFile.Url)"
+            DestinationFolder="$([System.IO.Path]::GetTempPath())/MLNET/"
+            Retries="5"
+            SkipUnchangedFiles="true">
+    </DownloadFile>
+  </Target>
 </Project>
