+++ test/Microsoft.ML.PerformanceTests/Numeric/Ranking.cs
@@ -11,7 +11,7 @@ using Microsoft.ML.Trainers.FastTree;
 using Microsoft.ML.Transforms;
 using Microsoft.ML.TestFrameworkCommon;
 
-namespace Microsoft.ML.Benchmarks
+namespace Microsoft.ML.PerformanceTests
 {
     [Config(typeof(TrainConfig))]
     public class RankingTrain : BenchmarkBase
