+++ test/Microsoft.ML.PerformanceTests/RffTransform.cs
@@ -4,12 +4,12 @@
 
 using System.IO;
 using BenchmarkDotNet.Attributes;
-using Microsoft.ML.Benchmarks.Harness;
+using Microsoft.ML.PerformanceTests.Harness;
 using Microsoft.ML.Data;
 using Microsoft.ML.TestFrameworkCommon;
 using Microsoft.ML.Transforms;
 
-namespace Microsoft.ML.Benchmarks
+namespace Microsoft.ML.PerformanceTests
 {
     [CIBenchmark]
     public class RffTransformTrain : BenchmarkBase
